convert double to int

+++ NextEld/Extensions/Double+Extension.swift
@@ -1,12 +1,12 @@
 
 
 extension Double {
-    func getHours() -> Double {
-        return self / 3600
+    func getHours() -> Int {
+        return Int(self / 3600)
     }
     
-    func getMin() -> Double {
-        return self / 60
+    func getMin() -> Int {
+        return Int(self / 60)
     }
     
 }
